backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.1
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings==2.9.1
python-jose==3.5.0
httpx==0.28.1
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
